refactor(dashboard-auth): move login views to auth module
