+++ apps/dashboard/src/modules/auth/components/BeerMugs.vue
