+++ apps/dashboard/src/modules/auth/routes.ts
@@ -1,0 +1,31 @@
+import type { RouteRecordRaw } from "vue-router";
+import AuthLoginView from "@/modules/auth/views/AuthLoginView.vue";
+import AuthLocalView from "@/modules/auth/views/AuthLocalView.vue";
+import AuthResetView from "@/modules/auth/views/AuthResetView.vue";
+import AuthLayout from "@/modules/auth/layouts/AuthLayout.vue";
+
+export function authRoutes(): RouteRecordRaw {
+  return {
+    path: '',
+    component: AuthLayout,
+    children: [
+      {
+        path: '',
+        component: AuthLoginView,
+        name: 'login',
+        alias: ['/login'],
+      },
+      {
+        path: '/local',
+        component: AuthLocalView,
+        name: 'local',
+        alias: ['/local'],
+      },
+      {
+        path: '/passwordreset',
+        component: AuthResetView,
+        name: 'passwordreset'
+      },
+    ],
+  };
+}
