+++ apps/dashboard/src/modules/auth/views/AuthLoginView.vue
@@ -1,13 +1,13 @@
 <template>
   <div>
-        <img class="max-h-9rem block mx-auto my-0" src="@/assets/img/bier.png" alt="logo"/>
+        <img class="max-h-9rem block mx-auto my-0" src="../../../assets/img/bier.png" alt="logo"/>
         <div class="text-900 text-5xl mt-0 mx-auto mb-2 w-full">{{ $t('login.SudoSOS Login') }}</div>
         <Button
             id="login-gewis-button"
             @click="loginViaGEWIS"
             class="my-3 mx-auto w-full flex justify-content-center align-items-center"
         >
-          <img class="mr-3 h-18" src="@/assets/img/gewis-branding.svg" alt="GEWIS"/>
+          <img class="mr-3 h-18" src="../../../assets/img/gewis-branding.svg" alt="GEWIS"/>
           {{ $t('login.Login via GEWIS') }}
         </Button>
         <Button
